refactor: Add delete functionality to EventDialog and EventRowComponent
Add delete functionality to the EventDialog and EventRowComponent components. This allows users to delete events from the table by clicking on a delete button. The handleDelete function is called when the delete button is clicked, and it sends a DELETE request to the server to remove the event. This improvement enhances the user experience by providing an easy way to delete events.